+++ configs/module/word.yaml
@@ -1,5 +1,5 @@
 cls:
-  _target_: kwja.models.word_module.WordModule
+  _target_: kwja.models.word.WordModule
 
 load_from_checkpoint:
   _target_: ${module.cls._target_}.load_from_checkpoint
